fix: COOP_TASKRUN event loop livelock and hash table infinite loop
Two critical bugs fixed:

1. Event loop livelock under high QoS 2 fan-out: intermediate
   ring_submit(0) calls during CQE processing created a feedback loop
   with COOP_TASKRUN where io_uring_enter generated new CQEs that kept
   the loop spinning. Fixed by submitting once at loop top, then draining
   all CQEs without any io_uring_enter during processing.

2. Infinite loop in inflight_hash_remove backward-shift deletion: the
   scanning cursor was recomputed from the gap pointer each iteration,
   so when should_shift was false the cursor never advanced. Fixed by
   using an independent scanning pointer that always moves forward.

Also includes: hot/cold inflight split for cache efficiency, async
PUBREL via submit_send_inf, sync sys_write fallback when resp_slots
exhausted, fast-path 4-byte protocol message parsing, send/recv retry
queues for SQ-full recovery, and SQPOLL removal in favor of
COOP_TASKRUN + SINGLE_ISSUER.

+++ src/broker/msg_path.h
@@ -80,7 +80,7 @@ struct broker;
 //
 // Returns: send_desc index, or -1 on failure
 INLINE i32 msg_prepare_send(struct broker *b, u32 msg_idx, u16 packet_id,
-                            u8 sub_qos, u8 dup, u32 slot_idx, u16 slot_gen) {
+                            u8 sub_qos, u8 dup, u32 slot_idx, u8 slot_gen) {
     struct canonical_msg *msg = msg_pool_get(&b->msg_pool, msg_idx);
     if (unlikely(!msg)) {
         return -1;
